Express module and static file server
Installed the Express module, which is a framework that is useful in
creating websites in a simple manner (the Express module belongs to the
Node Package Manager repository).
my_static_file_server.js creates a static file server. Anything put in
the /public folder can now be requested by the browser and displayed.
(e.g. if an image called my_image.png  is put inside the public folder,
it can be accessed by going to "http://localhost:8080/my_image.png".
This code was found on
http://blog.modulus.io/absolute-beginners-guide-to-nodejs