Code organization across files.
parser.js contains the parsing logic from the original my_parser.js
file. The "module.exports" line tells Node that you're exporting from
this file. In this case, it is exporting the constructor of the Parser
object, so that users can create instances of the Parser object.

my_parser.js is refactored to import and utilize the new Parser object.
Files are included the same way modules are, except a path is provided
instead of a name.