Websockets practice: finished
Wrote the html code myself, otherwise the rest of the code is from a
tutorial found here: http://codular.com/node-web-sockets

This is a bare-bones node.js / websockets application that displays a
chat-log across several clients and allows them all to update the
chat-log.

While this is very similar to what we are wanting to do for our
portfolio 2, I expect that we'll be tinkering with Express, CSS, and
other things to build on top of this concept/tutorial.

To run this: Run the terminal command "node myServer.js" while in the
correct directory. After the local websockets server is running, then
open the html web-page across multiple browsers so you can see the
multi-client push/pull of data at work.