Express/Jade/MongoDB Practice
This is the code that was created for a web app that utilizes the Express, Jade, and MongoDB modules of npm.
I used the code found in this tutorial: http://cwbuecheler.com/web/tutorials/2013/node-express-mongo/ and followed its steps to create the application.

To explain what all the tutorial covered in-depth would take more time than I have. I did have to install the Mongo client to be able to use the MongoDB module, and that isn't included in this commit.